@@ -1,17 +1,17 @@
 module.exports = {
   error(...args) {
-    console.error(...args)
+    console.error(...args);
   },
 
   warn(...args) {
-    console.warn(...args)
+    console.warn(...args);
   },
 
   info(...args) {
-    console.log(...args)
+    console.log(...args);
   },
 
   debug(...args) {
-    console.log(...args)
-  }
-}
+    console.log(...args);
+  },
+};
