@@ -1,67 +1,67 @@
-const zelcashrpc = require('zelcashrpc')
-const fullnode = require('fullnode')
+const zelcashrpc = require('zelcashrpc');
+const fullnode = require('fullnode');
 
 async function getInfo(req, res) {
-  const config = new fullnode.Config()
-  const rpcuser = config.rpcuser() || 'rpcuser'
-  const rpcpassword = config.rpcpassword() || 'rpcpassowrd'
-  const rpcport = config.rpcport() || 16124
+  const config = new fullnode.Config();
+  const rpcuser = config.rpcuser() || 'rpcuser';
+  const rpcpassword = config.rpcpassword() || 'rpcpassowrd';
+  const rpcport = config.rpcport() || 16124;
 
   const client = new zelcashrpc.Client({
     port: rpcport,
     user: rpcuser,
     pass: rpcpassword,
-    timeout: 60000
-  })
+    timeout: 60000,
+  });
 
   const response = {
     status: 'success',
-    data: 'Welcome to ZelNode network!'
-  }
+    data: 'Welcome to ZelNode network!',
+  };
 
   try {
-    const data = await client.getInfo()
-    response.status = 'success'
-    response.data = data
+    const data = await client.getInfo();
+    response.status = 'success';
+    response.data = data;
   } catch (err) {
-    response.status = 'error'
-    response.data = err
+    response.status = 'error';
+    response.data = err;
   }
 
-  return res.json(response)
+  return res.json(response);
 }
 
 async function getZelnNodeStatus(req, res) {
-  const config = new fullnode.Config()
-  const rpcuser = config.rpcuser() || 'rpcuser'
-  const rpcpassword = config.rpcpassword() || 'rpcpassowrd'
-  const rpcport = config.rpcport() || 16124
+  const config = new fullnode.Config();
+  const rpcuser = config.rpcuser() || 'rpcuser';
+  const rpcpassword = config.rpcpassword() || 'rpcpassowrd';
+  const rpcport = config.rpcport() || 16124;
 
   const client = new zelcashrpc.Client({
     port: rpcport,
     user: rpcuser,
     pass: rpcpassword,
-    timeout: 60000
-  })
+    timeout: 60000,
+  });
 
   const response = {
     status: 'success',
-    data: 'Welcome to ZelNode network!'
-  }
+    data: 'Welcome to ZelNode network!',
+  };
 
   try {
-    const data = await client.getzelnodestatus()
-    response.status = 'success'
-    response.data = data
+    const data = await client.getzelnodestatus();
+    response.status = 'success';
+    response.data = data;
   } catch (err) {
-    response.status = 'error'
-    response.data = err
+    response.status = 'error';
+    response.data = err;
   }
 
-  return res.json(response)
+  return res.json(response);
 }
 
 module.exports = {
   getInfo,
-  getZelnNodeStatus
-}
+  getZelnNodeStatus,
+};
