@@ -13,6 +13,6 @@
 <script>
 
 export default {
-  name: 'Header'
-}
+  name: 'Header',
+};
 </script>
