@@ -1,16 +1,17 @@
-import Vue from 'vue'
-import Router from 'vue-router'
-import MainPage from '@/components/MainPage'
+import Vue from 'vue';
+import Router from 'vue-router';
+import MainPage from '../components/MainPage';
 
-Vue.use(Router)
+Vue.use(Router);
 
 export default new Router({
   mode: 'history',
+  base: process.env.BASE_URL,
   routes: [
     {
       path: '/',
       name: 'MainPage',
-      component: MainPage
-    }
-  ]
-})
+      component: MainPage,
+    },
+  ],
+});
