@@ -1,13 +1,12 @@
 
-import axios from 'axios'
+import axios from 'axios';
 
-const config = require('../../../config/default')
-const userconfig = require('../../../config/userconfig')
-const port = config.server.localport
-const externalip = userconfig.initial.ipaddress
+const config = require('../../../config/default');
+const userconfig = require('../../../config/userconfig');
 
-export default () => {
-  return axios.create({
-    baseURL: `http://${externalip}:${port}`
-  })
-}
+const port = config.server.localport;
+const externalip = userconfig.initial.ipaddress;
+
+export default () => axios.create({
+  baseURL: `http://${externalip}:${port}`,
+});
