@@ -1,10 +1,10 @@
-import Api from '@/services/Api'
+import Api from '@/services/Api';
 
 export default {
   getInfo() {
-    return Api().get('/getinfo')
+    return Api().get('/getinfo');
   },
   getZelNodeStatus() {
-    return Api().get('/getzelnodestatus')
-  }
-}
+    return Api().get('/getzelnodestatus');
+  },
+};
