@@ -1,11 +1,11 @@
-import Api from '@/services/Api'
+import Api from '@/services/Api';
 
 export default {
   updateFlux(zelidauthHeader) {
     return Api().get('/zelnode/updateflux', {
       headers: {
-        zelidauth: zelidauthHeader
-      }
-    })
-  }
-}
+        zelidauth: zelidauthHeader,
+      },
+    });
+  },
+};
